Rework IP integration and board specific HDL
This commit seperates the uCaspian core HDL from external IP cores and
top-level wrappers for specific boards. This makes integrating uCaspian
into other designs and frameworks much easier (i.e. OpenLane, LiteX).

This also has the added advantage of making it easier to keep track of
what external IP is used. This change also removes the need to use
the legacy method of sharing models by using include headers.

The disadvantage of this approach is we have to list all the source
files of a design. This is typically done using a modern IP integration
framework like FuseSoC or LiteX.